Update #19 using export Service in for Metric view UI
LEX metric's files were edited to leave no location when computed for a
file. Also, the export Wizard is now common to both Violation & Metric
views. The service of the plugin fr.cnes.analysis.tools.export.xml was
improved to give a nameplace for method locations and only a TypePlace
for when the metric is intended to the whole file.
All test were updated to identify the CheckResult corresponding to the
old FileValue class (when no location is set).